new DB with Data

+++ src/main/java/by/vstu/vsrpp/exam/studgroup/repository/GroupRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import by.vstu.vsrpp.exam.studgroup.entity.Group;
 
-public interface GroupRepository extends JpaRepository<Group, Long> {
+public interface GroupRepository extends JpaRepository<Group, Integer> {
 
 }
